requirements-common.txt
# Common requirements used across local dev and production
FastAPI==0.65.2
gunicorn==20.0.4
starlette==0.14.2
uvicorn==0.11.7
# psycopg2==2.9.9
psycopg2-binary==2.9.9
